EPR - Ethnic Power Relations

The Ethnic Power Relations (EPR) Core Dataset 2019 is based on the original dataset EPR v.1.1 and identifies all politically relevant ethnic groups and their access to state power in every country of the world from 1946 to 2017. It includes annual data on over 800 groups and codes the degree to which their representatives held executive-level state power—from total control of the government to overt political discrimination.

epr_df

epr_status_codes_df

Format

Variables:

gwid

Gleditsch and Ward State ID

statename

from

to

group

groupid

gwgroupid

umbrella

size

status

The group rules alone:

"MONOPOLY"

Elite members hold monopoly power in the executive to the exclusion of members of all other ethnic groups.

"DOMINANCE"

Elite members of the group hold dominant power in the executive but there is some limited inclusion of “token” members of other groups who however do not have real influence on decision making.

The group shares power:

"SENIOR PARTNER"

Representatives of the group participate as senior partners in a formal or informal power-sharing arrangement. By power sharing, we mean any arrangement that divides executive power among leaders who claim to represent particular ethnic groups and who have real influence on political decision making.

"JUNIOR PARTNER"

Representatives participate as junior partners in government.

The choice between senior and junior partner depends on the group’s absolute influence in the executive - that is, irrespective of group size -, measured by the number and importance of the positions controlled by group members.

The group is excluded:

"POWERLESS"

Elite representatives hold no political power (or do not have influence on decision making) at the national level of executive power - although without being explicitly discriminated against.

"DISCRIMINATION"

Group members are subjected to active, intentional, and targeted discrimination by the state, with the intent of excluding them from political power. Such active discrimination can be either formal or informal, but always refers to the domain of public politics (excluding discrimination in the socio-economic sphere).

"SELF-EXCLUSION"

The special category of self-exclusion applies to groups that have excluded themselves from central state power, in the sense that they control a particular territory of the state which they have declared independent from the central government.

This category was labeled “separatist autonomy” in previous EPR versions.

reg_aut

Source

EPR - Ethnic Power Relations

Data (CSV)

Codebook

References

Vogt, Manuel, Nils-Christian Bormann, Seraina Rüegger, Lars-Erik Cederman, Philipp Hunziker, and Luc Girardin. 2015. “Integrating Data on Ethnicity, Geography, and Conflict: The Ethnic Power Relations Data Set Family.” Journal of Conflict Resolution 59(7): 1327–42.

Cederman, Lars-Erik, Andreas Wimmer, and Brian Min. 2010. “Why Do Ethnic Groups Rebel? New Data and Analysis.” World Politics 62(1): 87–119.

Examples

epr_df

#> # A tibble: 4,255 x 11
#>     gwid statename  from    to group groupid gwgroupid umbrella   size status
#>    <int> <chr>     <int> <int> <chr>   <int>     <int>    <int>  <dbl> <chr> 
#>  1     2 United S…  1946  1965 Whit…    1000    201000       NA 0.691  MONOP…
#>  2     2 United S…  1946  1965 Lati…    2000    202000       NA 0.125  IRREL…
#>  3     2 United S…  1946  1965 Afri…    3000    203000       NA 0.124  DISCR…
#>  4     2 United S…  1946  1965 Asia…    4000    204000       NA 0.036  IRREL…
#>  5     2 United S…  1946  1965 Amer…    5000    205000       NA 0.0078 POWER…
#>  6     2 United S…  1946  1965 Arab…    6000    206000       NA 0.0042 IRREL…
#>  7     2 United S…  1966  2008 Whit…    1000    201000       NA 0.691  DOMIN…
#>  8     2 United S…  1966  2008 Lati…    2000    202000       NA 0.125  POWER…
#>  9     2 United S…  1966  2008 Afri…    3000    203000       NA 0.124  POWER…
#> 10     2 United S…  1966  2008 Asia…    4000    204000       NA 0.036  POWER…
#> # … with 4,245 more rows, and 1 more variable: reg_aut <lgl>

epr_status_codes_df

#> # A tibble: 7 x 3
#>   status       category   definition                                            
#>   <chr>        <chr>      <chr>                                                 
#> 1 MONOPOLY     rules_alo… Elite members hold monopoly power in the executive to…
#> 2 DOMINANCE    rules_alo… Elite members of the group hold dominant power in the…
#> 3 SENIOR PART… shares_po… Representatives of the group participate as senior pa…
#> 4 JUNIOR PART… shares_po… Representatives participate as junior partners in gov…
#> 5 POWERLESS    is_exclud… Elite representatives hold no political power (or do …
#> 6 DISCRIMINAT… is_exclud… Group members are subjected to active, intentional, a…
#> 7 SELF-EXCLUS… is_exclud… The special category of self-exclusion applies to gro…